Rename AutosizeInput to Input
