+++ src/components/input/index.ts
@@ -1,2 +1,2 @@
-export { default as AutosizeInput } from './autosize-input';
+export { default as Input } from './input';
 export { default as TextareaInput } from './textarea-input';
